requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.20
chromadb==0.4.15
langchain==0.1.14
google-generativeai==0.3.2
python-dotenv==1.0.0